Drop code for dealing with the ls-R kpathsea database
Octave does not use the ls-R database anymore, so that its generation is
not need in the postinst script.  We still keep removing the file in the
prerm script for cleaning old installations of the package.  In
debian/rules, we still have to remove it after install, because the
upstream Makefile is not yet changed.  This is already changed upstream,
see
http://www-old.cae.wisc.edu/pipermail/octave-maintainers/2009-February/010708.html.