d/octave.triggers: Use a -noawait type of trigger.
So far the trigger for Octave Forge packages was implicitly of the -await type.
Lintian now complains about this, and asks to explicitly make a choice between
the -await and -noawait types (the latter being easier to manage for dpkg).
Since Octave Forge packages can be configured before the trigger is activated,
we go for the -noawait type.

Gbp-Dch: Full